@@ -1,3 +1,3 @@
 module github.com/go-task/template
 
-go 1.21.0
+go 1.23.0
